requirements.txt
dacite
python-dateutil
httpx